rental-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
PyJWT
psycopg2-binary==2.9.9
